fix: change fileSize from primitive long to Long wrapper
Primitive long defaults to 0 which bypassed validation (<= 0 check).
Now using Long wrapper so null/missing values fail validation.
Updated FileStorageDomainService and tests to use (long) casts.

+++ src/main/java/com/wpanther/storage/application/usecase/FileStorageDomainService.java
@@ -68,7 +68,7 @@ public class FileStorageDomainService implements DocumentStorageUseCase {
                 .contentType(contentType != null ? contentType : ContentTypeUtil.determineContentType(filename))
                 .storagePath(result.location())
                 .storageUrl(result.location())
-                .fileSize(content.length)
+                .fileSize((long) content.length)
                 .checksum(checksum)
                 .createdAt(LocalDateTime.now())
                 .build();
